+++ src/core/embedders/FontFlags.ts
@@ -35,11 +35,11 @@ const makeFontFlags = (options: FontFlagOptions) => {
 export const deriveFontFlags = (font: Font): number => {
   const familyClass = font['OS/2'] ? font['OS/2'].sFamilyClass : 0;
   const flags = makeFontFlags({
-    fixedPitch: font.post.isFixedPitch,
+    fixedPitch: !!font.post?.isFixedPitch,
     serif: 1 <= familyClass && familyClass <= 7,
     symbolic: true, // Assume the font uses non-latin characters
     script: familyClass === 10,
-    italic: font.head.macStyle.italic,
+    italic: !!font.head?.macStyle?.italic,
   });
   return flags;
 };